refactor(api): restore injected route dependencies
