--- api/lib/serialPorts.ts
@@ -1,4 +1,0 @@
-import { Driver } from 'zwave-js'
-
-export const enumerateSerialPorts: typeof Driver.enumerateSerialPorts =
-	Driver.enumerateSerialPorts.bind(Driver)
