+++ package.json
@@ -1,9 +1,6 @@
 {
   "name": "zwave-js-ui",
   "type": "module",
-  "imports": {
-    "#api/*": "./api/*"
-  },
   "version": "11.21.1",
   "description": "Z-Wave Control Panel and MQTT Gateway",
   "keywords": [
